silence a size conversion warning
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@18776 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/Utils/PathUtils.cpp
@@ -431,7 +431,7 @@ CString CPathUtils::PathUnescape (const char* path)
 
 	// no escapement necessary, just unicode conversion
 	CString result;
-	CUnicodeUtils::UTF8ToUTF16 (path, i+1, result.GetBufferSetLength (i+1));
+	CUnicodeUtils::UTF8ToUTF16 (path, i+1, result.GetBufferSetLength ((int)i+1));
 	result.ReleaseBuffer();
 
 	return result;
